requirements.txt
flask==2.0.1
google-generativeai==0.3.2
langchain-community==0.0.27
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
gunicorn==20.1.0
SpeechRecognition==3.10.1
gTTS==2.5.1
pygame==2.5.2 